% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{deckglOutput}
\alias{deckglOutput}
\title{Shiny output for Deck.gl widget}
\usage{
deckglOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable name}

\item{width, height}{CSS dimensions (e.g. '100%', '400px') for the container.}
}
\value{
A Shiny output binding for use in the UI.
}
\description{
Use this in the UI to create a placeholder for the Deck.gl visualization.
}
\examples{
if (interactive()) {
  library(shiny)
  library(rDeckgl)

  ui <- fluidPage(
    deckglOutput("myDeckgl", width = "100\%", height = "600px")
  )

  server <- function(input, output, session) {
    output$myDeckgl <- renderDeckgl({
      deckgl(spec = my_spec, data = my_data)
    })
  }

  shinyApp(ui, server)
}

}
