% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl.R
\name{hydrate_deckgl_spec}
\alias{hydrate_deckgl_spec}
\title{Hydrate Deck.gl DuckDB data references}
\usage{
hydrate_deckgl_spec(spec, con)
}
\arguments{
\item{spec}{The Deck.gl spec list}

\item{con}{The DuckDB connection}
}
\value{
A spec list where DuckDB nodes are replaced with Arrow IPC payloads
}
\description{
Hydrate Deck.gl DuckDB data references
}
\keyword{internal}
