% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDeckgl-package.R
\docType{package}
\name{rDeckgl-package}
\alias{rDeckgl}
\alias{rDeckgl-package}
\title{rDeckgl: Deck.gl for R}
\description{
Provides R bindings for deck.gl, a WebGL framework for rendering large interactive spatial and tabular visualizations. The package supplies htmlwidgets and Shiny bindings, supports DuckDB-backed data hydration, and bundles the JavaScript assets needed to render deck.gl specifications from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TiRizvanov/rDeckgl}
  \item \url{https://tirizvanov.github.io/rDeckgl/}
  \item Report bugs at \url{https://github.com/TiRizvanov/rDeckgl/issues}
}

}
\author{
\strong{Maintainer}: Timur Rizvanov \email{timurr@bu.edu}

Authors:
\itemize{
  \item Edward C. Ruiz \email{ecr7407@gmail.com}
  \item Ruben Dries \email{rubendries@gmail.com} [reviewer]
}

}
\keyword{internal}
