% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDeckgl-package.R
\docType{package}
\name{rDeckgl-package}
\alias{rDeckgl}
\alias{rDeckgl-package}
\title{rDeckgl: Deck.gl for R}
\description{
Provides an R interface to 'deck.gl', a WebGL-powered framework for visual exploratory data analysis of large datasets. Enables high-performance rendering of Arrow IPC streams directly from 'DuckDB' to the browser.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TiRizvanov/rDeckgl}
  \item Report bugs at \url{https://github.com/TiRizvanov/rDeckgl/issues}
}

}
\author{
\strong{Maintainer}: Timur Rizvanov \email{timurr@bu.edu}

Authors:
\itemize{
  \item Edward C. Ruiz \email{ecr7407@gmail.com}
  \item Ruben Dries \email{rubendries@gmail.com} [reviewer]
}

}
\keyword{internal}
