% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDeckgl-package.R
\docType{package}
\name{rDeckgl-package}
\alias{rDeckgl}
\alias{rDeckgl-package}
\title{rDeckgl: Deck.gl for R}
\description{
Provides R bindings for deck.gl, a WebGL framework for rendering large interactive spatial and tabular visualizations. The package supplies htmlwidgets and Shiny bindings, supports DuckDB-backed data hydration, and bundles the JavaScript assets needed to render deck.gl specifications from R.
}
\section{Acknowledgements}{

Developed in the Dries Lab at Boston University. This work was supported by
the Dries Lab and the Boston University Undergraduate Research Opportunities
Program (UROP).
}

\author{
\strong{Maintainer}: Timur Rizvanov \email{timurr@bu.edu}

Authors:
\itemize{
  \item Edward C. Ruiz \email{ecr7407@gmail.com}
  \item Ruben Dries \email{rubendries@gmail.com} [reviewer]
}

Other contributors:
\itemize{
  \item Dries Lab (Host laboratory and funding source) [funder]
  \item Boston University (Undergraduate Research Opportunities Program (UROP)) [funder]
}

}
\keyword{internal}
