% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDeckgl-package.R
\docType{package}
\name{rDeckgl-package}
\alias{rDeckgl}
\alias{rDeckgl-package}
\title{rDeckgl: Deck.gl for R}
\description{
R htmlwidget bindings for deck.gl, a WebGL-powered framework for visual exploratory data analysis of large datasets. Create interactive visualizations using deck.gl JSON specifications with server-side data hydration via DuckDB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TiRizvanov/rDeckgl}
  \item Report bugs at \url{https://github.com/TiRizvanov/rDeckgl/issues}
}

}
\author{
\strong{Maintainer}: Timur Rizvanov \email{timurr@bu.edu}

}
\keyword{internal}
